+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/service/UserDetailsService.kt
@@ -6,17 +6,15 @@ import com.saveourtool.save.domain.Role
 import com.saveourtool.save.domain.UserSaveStatus
 import com.saveourtool.save.entities.OriginalLogin
 import com.saveourtool.save.entities.User
-import com.saveourtool.save.utils.IdentitySourceAwareUserDetails
+import com.saveourtool.save.utils.getIdentitySourceAwareUserDetails
 import com.saveourtool.save.utils.orNotFound
 import org.slf4j.LoggerFactory
 import org.springframework.security.core.Authentication
 import org.springframework.security.core.userdetails.ReactiveUserDetailsService
 import org.springframework.security.core.userdetails.UserDetails
-import org.springframework.security.core.userdetails.UsernameNotFoundException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import reactor.core.publisher.Mono
-import reactor.kotlin.core.publisher.switchIfEmpty
 import reactor.kotlin.core.publisher.toMono
 import java.util.*
 
@@ -58,29 +56,6 @@ class UserDetailsService(
             }
         user.let { userRepository.save(it) }
     }
-
-    private fun Mono<User>.getIdentitySourceAwareUserDetails(username: String, source: String? = null) = this
-        .map<UserDetails> { user ->
-            user.toIdentitySourceAwareUserDetails()
-        }
-        .switchIfEmpty {
-            Mono.fromCallable {
-                val sourceMsg = source?.let {
-                    " and source=$source"
-                }.orEmpty()
-                logger.warn("Couldn't find user with name=$username$sourceMsg in DB!")
-            }.flatMap {
-                Mono.error(UsernameNotFoundException(username))
-            }
-        }
-    @Suppress("UnsafeCallOnNullableType")
-    private fun User.toIdentitySourceAwareUserDetails(): IdentitySourceAwareUserDetails = IdentitySourceAwareUserDetails(
-        username = this.name!!,
-        password = this.password.orEmpty(),
-        authorities = this.role,
-        identitySource = this.source,
-        id = this.id!!,
-    )
 
     /**
      * @param authentication
