+++ save-cloud-common/src/jvmMain/kotlin/com/saveourtool/save/utils/IdentitySourceAwareUserDetails.kt
@@ -1,7 +1,13 @@
 package com.saveourtool.save.utils
 
+import org.slf4j.LoggerFactory
 import org.springframework.security.core.authority.SimpleGrantedAuthority
-import org.springframework.security.core.userdetails.User
+import org.springframework.security.core.userdetails.UserDetails
+import org.springframework.security.core.userdetails.UsernameNotFoundException
+import reactor.core.publisher.Mono
+import reactor.kotlin.core.publisher.switchIfEmpty
+
+private val logger = LoggerFactory.getLogger(object {}.javaClass.enclosingClass::class.java)
 
 /**
  * @param authorities comma-separated authorities as a single string
@@ -14,7 +20,7 @@ class IdentitySourceAwareUserDetails(
     authorities: String?,
     val identitySource: String,
     val id: Long,
-) : User(
+) : org.springframework.security.core.userdetails.User(
     username,
     password,
     authorities?.split(',')
@@ -22,3 +28,34 @@ class IdentitySourceAwareUserDetails(
         ?.map { SimpleGrantedAuthority(it) }
         .orEmpty()
 )
+
+/**
+ * @return IdentitySourceAwareUserDetails, retrieved from save-cloud User entity
+ */
+fun com.saveourtool.save.entities.User.toIdentitySourceAwareUserDetails(): IdentitySourceAwareUserDetails = IdentitySourceAwareUserDetails(
+    username = this.name!!,
+    password = this.password.orEmpty(),
+    authorities = this.role,
+    identitySource = this.source,
+    id = this.id!!,
+)
+
+/**
+ * @param username
+ * @param source
+ * @return mono of IdentitySourceAwareUserDetails, retrieved from save-cloud User entity
+ */
+fun Mono<com.saveourtool.save.entities.User>.getIdentitySourceAwareUserDetails(username: String, source: String? = null) = this
+    .map<UserDetails> { user ->
+        user.toIdentitySourceAwareUserDetails()
+    }
+    .switchIfEmpty {
+        Mono.fromCallable {
+            val sourceMsg = source?.let {
+                " and source=$source"
+            }.orEmpty()
+            logger.warn("Couldn't find user with name=$username$sourceMsg in DB!")
+        }.flatMap {
+            Mono.error(UsernameNotFoundException(username))
+        }
+    }
